requirements.txt
pyspark
kafka-python
six
praw